requirements.txt
antlr4-python3-runtime==4.7.1
datrie>=0.8.2
grammar_graph>=0.1.0
isla-solver>=0.8.0
pathos>=0.2.9
setuptools-antlr>=0.4.0
toml>=0.10.2
z3-solver>=4.10.2.0
